requirements.txt
flask
opencv-python
tensorflow
numpy
pillow
streamlit 